+++ src/react/components/chart/customElements/ChartScaffold.js
@@ -201,7 +201,7 @@ class ChartScaffold extends PureComponent {
     const {
       qaHook,
       // Path to an image to be used with avatar label.
-      avatarImageUrl,
+      avatarImageUrlData,
 
       // Chart...
       chartId,
@@ -491,7 +491,7 @@ class ChartScaffold extends PureComponent {
                         labelHeight={chartBottom}
                         isToolTipHidden={isXAxisToolTipHidden}
                         updateToolTip={this.updateToolTip}
-                        avatarImageUrl={avatarImageUrl}
+                        avatarImageUrlData={avatarImageUrlData}
                         barsData={barsData}
                       />
                     }
@@ -578,5 +578,5 @@ ChartScaffold.propTypes = {
   paginationNextTitle: PropTypes.string,
   paginationPreviousTitle: PropTypes.string,
   createPaginationMessage: PropTypes.func,
-  avatarImageUrl: PropTypes.array,
+  avatarImageUrlData: PropTypes.array,
 };